Revert "update code structure"
This reverts commit 693978bda63cb436f92646990b175a92a9619304.
