+++ src/lotus/client/mod.rs
@@ -3,8 +3,19 @@
 mod ipc;
 mod lotus;
 
+use anyhow::Result;
+use cid::Cid;
+use fvm_shared::econ::TokenAmount;
+use num_traits::cast::ToPrimitive;
+use serde_json::json;
+
 use crate::config::Subnet;
 use crate::jsonrpc::{JsonRpcClient, JsonRpcClientImpl};
+use crate::lotus::message::mpool::{
+    MpoolPushMessage, MpoolPushMessageResponse, MpoolPushMessageResponseInner,
+};
+use crate::lotus::message::CIDMap;
+use crate::lotus::StateWaitMsgResponse;
 
 // RPC methods
 mod methods {
@@ -50,6 +61,69 @@ pub struct LotusJsonRPCClient<T: JsonRpcClient> {
 impl<T: JsonRpcClient> LotusJsonRPCClient<T> {
     pub fn new(client: T) -> Self {
         Self { client }
+    }
+}
+
+impl<T: JsonRpcClient + Send + Sync> LotusJsonRPCClient<T> {
+    async fn mpool_push_message_inner(
+        &self,
+        msg: MpoolPushMessage,
+    ) -> Result<MpoolPushMessageResponseInner> {
+        let nonce = msg
+            .nonce
+            .map(|n| serde_json::Value::Number(n.into()))
+            .unwrap_or(serde_json::Value::Null);
+
+        let to_value = |t: Option<TokenAmount>| {
+            t.map(|n| serde_json::Value::Number(n.atto().to_u64().unwrap().into()))
+                .unwrap_or(serde_json::Value::Null)
+        };
+        let gas_limit = to_value(msg.gas_limit);
+        let gas_premium = to_value(msg.gas_premium);
+        let gas_fee_cap = to_value(msg.gas_fee_cap);
+        let max_fee = to_value(msg.max_fee);
+
+        // refer to: https://lotus.filecoin.io/reference/lotus/mpool/#mpoolpushmessage
+        let params = json!([
+            {
+                "to": msg.to.to_string(),
+                "from": msg.from.to_string(),
+                "value": msg.value.atto().to_string(),
+                "method": msg.method,
+                "params": msg.params,
+
+                // THESE ALL WILL AUTO POPULATE if null
+                "nonce": nonce,
+                "gas_limit": gas_limit,
+                "gas_fee_cap": gas_fee_cap,
+                "gas_premium": gas_premium,
+                "cid": CIDMap::from(msg.cid),
+                "version": serde_json::Value::Null,
+            },
+            {
+                "max_fee": max_fee
+            }
+        ]);
+
+        let r = self
+            .client
+            .request::<MpoolPushMessageResponse>(methods::MPOOL_PUSH_MESSAGE, params)
+            .await?;
+        log::debug!("received mpool_push_message response: {r:?}");
+
+        Ok(r.message)
+    }
+
+    async fn state_wait_msg_inner(&self, cid: Cid, nonce: u64) -> Result<StateWaitMsgResponse> {
+        // refer to: https://lotus.filecoin.io/reference/lotus/state/#statewaitmsg
+        let params = json!([CIDMap::from(cid), nonce]);
+
+        let r = self
+            .client
+            .request::<StateWaitMsgResponse>(methods::STATE_WAIT_MSG, params)
+            .await?;
+        log::debug!("received state_wait_msg response: {r:?}");
+        Ok(r)
     }
 }
